added parsing the response, moved mapping code to a separate service, added tests; updated main application code

+++ src/main/java/com/spond/WeatherService/domain/WeatherForecast.java
@@ -2,5 +2,5 @@ package com.spond.WeatherService.domain;
 
 import java.time.LocalDateTime;
 
-public record WeatherForecast(int temperature, double windSpeed, LocalDateTime lastUpdated) {
+public record WeatherForecast(double temperature, double windSpeed, LocalDateTime lastUpdated) {
 }
